Встроена новая версия библиотеки

+++ VideoExample/app/src/main/java/com/example/anya/videoexample/SimpleActivity.java
@@ -2,41 +2,18 @@ package com.example.anya.videoexample;
 
 import android.app.Activity;
 import android.content.res.AssetManager;
-import android.content.res.Configuration;
-import android.net.Uri;
 import android.os.Bundle;
 import android.util.Log;
-import android.view.Gravity;
-import android.view.SurfaceHolder;
-import android.view.SurfaceView;
-import android.view.ViewGroup.LayoutParams;
-import android.widget.Toast;
-import org.videolan.libvlc.IVLCVout;
-import org.videolan.libvlc.LibVLC;
-import org.videolan.libvlc.Media;
-import org.videolan.libvlc.MediaPlayer;
+import android.widget.FrameLayout;
 
 import java.io.*;
-import java.lang.ref.WeakReference;
-import java.util.ArrayList;
 
-public class SimpleActivity extends Activity implements IVLCVout.Callback {
+public class SimpleActivity extends Activity {
     public final static String TAG = "LibVLCAndroidSample/VideoActivity";
-
-    public final static String LOCATION = "com.compdigitec.libvlcandroidsample.VideoActivity.location";
 
     private String mFilePath;
 
-    // display surface
-    private SurfaceView mSurface;
-    private SurfaceHolder holder;
-
-    // media player
-    private LibVLC libvlc;
-    private MediaPlayer mMediaPlayer = null;
-    private int mVideoWidth;
-    private int mVideoHeight;
-    private final static int VideoSizeChanged = -1;
+    private VideoViewVLC mSurface;
 
     File mFile;
 
@@ -49,35 +26,33 @@ public class SimpleActivity extends Activity implements IVLCVout.Callback {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
 
-        // Receive path to play from intent
-        //Intent intent = getIntent();
-        ///mFilePath = intent.getExtras().getString(LOCATION);
-        //mFilePath = "http://media5.krasview.ru/download/5970c1f0ce53d76.mp4";
         AssetManager myAssetManager = getApplicationContext().getAssets();
 
-        InputStream in = null;
-        OutputStream out = null;
         String filename = "radioastron.mp4";
-        try {
-            in = myAssetManager.open(filename);
-            File outFile = new File(getExternalFilesDir(null), filename);
-            out = new FileOutputStream(outFile);
-            copyFile(in, out);
-        } catch(IOException e) {
-            Log.e("tag", "Failed to copy asset file: " + filename, e);
-        }
 
-        //mFilePath = "file:///android_asset/radioastron.mp4";
         mFile = new File(getExternalFilesDir(null), filename);
         mFilePath = mFile.getPath();
+
+        if(!mFile.exists()) {
+            InputStream in = null;
+            OutputStream out = null;
+            try {
+                in = myAssetManager.open(filename);
+                File outFile = new File(getExternalFilesDir(null), filename);
+                out = new FileOutputStream(outFile);
+                copyFile(in, out);
+            } catch(IOException e) {
+                Log.e("tag", "Failed to copy asset file: " + filename, e);
+            }
+        }
+
         Log.d("MyVLC", "exists" + mFile.exists());
 
-        //Log.d(TAG, "Playing back " + mFilePath);
+        FrameLayout frame = (FrameLayout) findViewById(R.id.player_surface_frame);
 
-
-        mSurface = (SurfaceView) findViewById(R.id.player_surface);
-        holder = mSurface.getHolder();
-        //holder.addCallback(this);
+        mSurface = new VideoViewVLC(this);
+        mSurface.setLayoutParams(new FrameLayout.LayoutParams(400, 300));
+        frame.addView(mSurface);
     }
 
     private void copyFile(InputStream in, OutputStream out) throws IOException {
@@ -88,198 +63,23 @@ public class SimpleActivity extends Activity implements IVLCVout.Callback {
         }
     }
 
-    @Override
-    public void onConfigurationChanged(Configuration newConfig) {
-        super.onConfigurationChanged(newConfig);
-        setSize(mVideoWidth, mVideoHeight);
-    }
 
     @Override
     protected void onResume() {
         super.onResume();
-        createPlayer(mFilePath);
+        mSurface.setVideoAndStart("http://media5.krasview.ru/download/354c6945d7af5e5.mp4");
     }
 
     @Override
     protected void onPause() {
         super.onPause();
-        releasePlayer();
+        mSurface.releasePlayer();
     }
 
     @Override
     protected void onDestroy() {
         super.onDestroy();
-        releasePlayer();
+        mSurface.releasePlayer();
     }
-
-    /*************
-     * Surface
-     *************/
-    private void setSize(int width, int height) {
-        mVideoWidth = width;
-        mVideoHeight = height;
-        if (mVideoWidth * mVideoHeight <= 1)
-            return;
-
-        if(holder == null || mSurface == null)
-            return;
-
-        // get screen size
-        int w = getWindow().getDecorView().getWidth();
-        int h = getWindow().getDecorView().getHeight();
-
-        // getWindow().getDecorView() doesn't always take orientation into
-        // account, we have to correct the values
-        boolean isPortrait = getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT;
-        if (w > h && isPortrait || w < h && !isPortrait) {
-            int i = w;
-            w = h;
-            h = i;
-        }
-
-        float videoAR = (float) mVideoWidth / (float) mVideoHeight;
-        float screenAR = (float) w / (float) h;
-
-        if (screenAR < videoAR)
-            h = (int) (w / videoAR);
-        else
-            w = (int) (h * videoAR);
-
-        // force surface buffer size
-        holder.setFixedSize(mVideoWidth, mVideoHeight);
-
-        // set display size
-        LayoutParams lp = mSurface.getLayoutParams();
-        lp.width = w;
-        lp.height = h;
-        mSurface.setLayoutParams(lp);
-        mSurface.invalidate();
-    }
-
-    /*************
-     * Player
-     *************/
-
-    private void createPlayer(String media) {
-        releasePlayer();
-        try {
-            if (media.toString().length() > 0) {
-                Toast toast = Toast.makeText(this, media.toString(), Toast.LENGTH_LONG);
-                toast.setGravity(Gravity.BOTTOM | Gravity.CENTER_HORIZONTAL, 0,
-                        0);
-                toast.show();
-            }
-
-            // Create LibVLC
-            // TODO: make this more robust, and sync with audio demo
-            ArrayList<String> options = new ArrayList<String>();
-            //options.add("--subsdec-encoding <encoding>");
-            options.add("--aout=opensles");
-            options.add("--audio-time-stretch"); // time stretching
-            options.add("-vvv"); // verbosity
-            libvlc = new LibVLC(options);
-            //libvlc.setOnHardwareAccelerationError(this);
-            holder.setKeepScreenOn(true);
-
-            // Create media player
-            mMediaPlayer = new MediaPlayer(libvlc);
-            mMediaPlayer.setEventListener(mPlayerListener);
-
-            // Set up video output
-            final IVLCVout vout = mMediaPlayer.getVLCVout();
-            vout.setVideoView(mSurface);
-            //vout.setSubtitlesView(mSurfaceSubtitles);
-            vout.addCallback(this);
-            vout.attachViews();
-            Uri uri = Uri.parse(media);
-            Log.d("MyVLC", "" + uri);
-            Media m = new Media(libvlc, mFilePath);
-            mMediaPlayer.setMedia(m);
-            mMediaPlayer.play();
-        } catch (Exception e) {
-            Toast.makeText(this, "Error creating player!", Toast.LENGTH_LONG).show();
-        }
-    }
-
-    // TODO: handle this cleaner
-    private void releasePlayer() {
-        if (libvlc == null)
-            return;
-        mMediaPlayer.stop();
-        final IVLCVout vout = mMediaPlayer.getVLCVout();
-        vout.removeCallback(this);
-        vout.detachViews();
-        holder = null;
-        libvlc.release();
-        libvlc = null;
-
-        mVideoWidth = 0;
-        mVideoHeight = 0;
-    }
-
-    /*************
-     * Events
-     *************/
-
-    private MediaPlayer.EventListener mPlayerListener = new MyPlayerListener(this);
-
-    @Override
-    public void onNewLayout(IVLCVout vout, int width, int height, int visibleWidth, int visibleHeight, int sarNum, int sarDen) {
-        if (width * height == 0)
-            return;
-
-        // store video size
-        mVideoWidth = width;
-        mVideoHeight = height;
-        setSize(mVideoWidth, mVideoHeight);
-    }
-
-    @Override
-    public void onSurfacesCreated(IVLCVout vout) {
-
-    }
-
-    @Override
-    public void onSurfacesDestroyed(IVLCVout vout) {
-
-    }
-
-    @Override
-    public void onHardwareAccelerationError(IVLCVout vlcVout) {
-
-    }
-
-    private static class MyPlayerListener implements MediaPlayer.EventListener {
-        private WeakReference<SimpleActivity> mOwner;
-
-        public MyPlayerListener(SimpleActivity owner) {
-            mOwner = new WeakReference<SimpleActivity>(owner);
-        }
-
-        @Override
-        public void onEvent(MediaPlayer.Event event) {
-            SimpleActivity player = mOwner.get();
-
-            switch(event.type) {
-                case MediaPlayer.Event.EndReached:
-                    Log.d(TAG, "MediaPlayerEndReached");
-                    player.releasePlayer();
-                    break;
-                case MediaPlayer.Event.Playing:
-                case MediaPlayer.Event.Paused:
-                case MediaPlayer.Event.Stopped:
-                default:
-                    break;
-            }
-        }
-    }
-
-   /* @Override
-    public void eventHardwareAccelerationError() {
-        // Handle errors with hardware acceleration
-        Log.e(TAG, "Error with hardware acceleration");
-        this.releasePlayer();
-        Toast.makeText(this, "Error with hardware acceleration", Toast.LENGTH_LONG).show();
-    }*/
 }
